Added infrastructure for RC builds
Added a `RC_build` property in the build script to make RC builds
for testing purposes.

The RC build use a different package_index file called
package_index_staging.json.

When the IDE detects the presence of the staging index in his
dist/ folder then this file is used instead of the main public
index to allow testing of new AVR-bundled cores.
Updates are also downloaded from a different URL and the original
package_index is ignored and left untouched.